refactor: move template renderer to its own file
no functional changes
